# flask/app/static/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Access API",
        "description": "API endpoints for UGV-Based Disaster Management System",
        "version": "1.0"
    },
    "basePath": "/",
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header",
            "description": "JWT Bearer token"
        }
    },
    "paths": {
        "/api/users/signup": {
            "post": {
                "tags": [
                    "User Management"
                ],
                "summary": "User Signup",
                "description": "Register a new user\n\npassword conditions:\n-contains at least 8 characters\n-contains at least one uppercase letter, one lowercase letter, one digit, and one special character\n\nnote: email and username must be unique",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "User data",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "email": {
                                    "type": "string",
                                    "example": "user@example.com",
                                    "required": true
                                },
                                "password": {
                                    "type": "string",
                                    "example": "Password@123",
                                    "required": true
                                },
                                "username": {
                                    "type": "string",
                                    "example": "user123",
                                    "required": true
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "User created successfully",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "User created successfully. Once you receive approval from the admin, you can access your account and establish a connection with the Cloud MQTT broker using the same credentials."
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "409": {
                        "description": "Conflict"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/users/login": {
            "post": {
                "tags": [
                    "User Management"
                ],
                "summary": "User Login",
                "description": "Login with email or username and password",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "User data",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "email_or_username": {
                                    "type": "string",
                                    "example": "user@example.com",
                                    "required": true
                                },
                                "password": {
                                    "type": "string",
                                    "example": "New_password123",
                                    "required": true
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User logged in successfully",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "User layla21 loggedin successfully."
                                },
                                "username": {
                                    "type": "string",
                                    "example": "layla21"
                                },
                                "type": {
                                    "type": "int",
                                    "example": 1
                                },
                                "token": {
                                    "type": "string",
                                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNjA1NWEwYWU4MGU5MGUwODY0MWUyMmVmIiwiaWF0IjoxNjYxNDM3NTQzLCJleHAiOjE2NjE0NDEyNDN9.T6PlvZxId2GJWnT5SyZgnkr0-pUgBp9_1l1dUKhxO54"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized - Invalid email or password"
                    },
                    "403": {
                        "description": "Forbidden Error"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/users/rtmp_auth": {
            "get": {
                "tags": [
                    "User Management"
                ],
                "summary": "Authentication Proxy for RTMP Server",
                "description": "Login with username and password for RTMP Server",
                "parameters": [
                    {
                        "in": "query",
                        "name": "username",
                        "type": "string",
                        "required": true,
                        "default": "user123"
                    },
                    {
                        "in": "query",
                        "name": "password",
                        "type": "string",
                        "required": true,
                        "default": "New_password123"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User logged in successfully"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized - Invalid email or password"
                    },
                    "403": {
                        "description": "Forbidden Error"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/api/users/logout": {
            "post": {
                "tags": [
                    "User Management"
                ],
                "summary": "User Logout",
                "description": "Logout the user",
                "responses": {
                    "200": {
                        "description": "User logged out successfully",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "User logged out successfully."
                                },
                                "token": {
                                    "type": "string",
                                    "example": ""
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid token"
                    },
                    "404": {
                        "description": "Not Found - User not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/users": {
            "get": {
                "tags": [
                    "User Management"
                ],
                "summary": "User Info Retrievement",
                "description": "Get User Information",
                "responses": {
                    "200": {
                        "description": "JSON object of user information",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "user_id": {
                                    "type": "string",
                                    "example": "6055a0ae80e90e08641e22ef"
                                },
                                "email": {
                                    "type": "string",
                                    "example": "email@example.com"
                                },
                                "username": {
                                    "type": "string",
                                    "example": "Username1"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid token"
                    },
                    "404": {
                        "description": "Not Found - User not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User Management"
                ],
                "summary": "User Info Update",
                "description": "Update user's username or email",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "User data",
                        "required": false,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "username": {
                                    "type": "string",
                                    "example": "new_username"
                                },
                                "email": {
                                    "type": "string",
                                    "example": "new_email@example.com"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User information updated successfully",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "User information is updated successfully."
                                },
                                "email": {
                                    "type": "string",
                                    "example": "email@example.com"
                                },
                                "username": {
                                    "type": "string",
                                    "example": "Username1"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found - User not found"
                    },
                    "409": {
                        "description": "Conflict - The username or email is already taken or identical to the current one"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/users/{user_id}": {
            "get": {
                "tags": [
                    "User Management"
                ],
                "summary": "User Info Retrievement [ADMIN]",
                "description": "Get User Information for Admin interface",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "ID of the user to retrieve",
                        "required": true,
                        "type": "string",
                        "example": "6055a0ae80e90e08641e22ef"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "JSON object of user information",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "user_id": {
                                    "type": "string",
                                    "example": "6055a0ae80e90e08641e22ef"
                                },
                                "email": {
                                    "type": "string",
                                    "example": "email@example.com"
                                },
                                "username": {
                                    "type": "string",
                                    "example": "Username1"
                                },
                                "type": {
                                    "type": "int",
                                    "example": 1
                                },
                                "status": {
                                    "type": "int",
                                    "example": 1
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid token"
                    },
                    "404": {
                        "description": "Not Found - User not found"
                    },
                    "403": {
                        "description": "Forbidden Error"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User Management"
                ],
                "summary": "User Deletion [ADMIN]",
                "description": "Delete User for Admin interface",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "ID of the user to delete",
                        "required": true,
                        "type": "string",
                        "example": "6055a0ae80e90e08641e22ef"
                    }
                ],
                "responses": {
                    "200": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "User is deleted successfully."
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden Error"
                    },
                    "404": {
                        "description": "Not Found - User not found"
                    },
                    "409": {
                        "description": "Conflict - User is already Inactive."
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/users/all": {
            "get": {
                "tags": [
                    "User Management"
                ],
                "summary": "User List Retrievement [ADMIN]",
                "description": "Get list of users with pagenation for Admin interface",
                "parameters": [
                    {
                        "in": "query",
                        "name": "page-number",
                        "type": "integer",
                        "required": false,
                        "description": "Page number",
                        "default": 1
                    },
                    {
                        "in": "query",
                        "name": "page-size",
                        "type": "integer",
                        "required": false,
                        "description": "Number of items per page",
                        "default": 5
                    },
                    {
                        "in": "query",
                        "name": "username",
                        "type": "string",
                        "required": false,
                        "description": "filter for username",
                        "default": "user"
                    },
                    {
                        "in": "query",
                        "name": "status",
                        "required": false,
                        "description": "filter for user status",
                        "type": "array",
                        "collectionFormat": "multi",
                        "items": {
                            "type": "integer"
                        }
                    },
                    {
                        "in": "query",
                        "name": "type",
                        "type": "array",
                        "required": false,
                        "description": "filter for user type",
                        "collectionFormat": "multi",
                        "items": {
                            "type": "integer"
                        }
                    },
                    {
                        "in": "query",
                        "name": "mission",
                        "type": "string",
                        "required": false,
                        "description": "[mission_id] users for mission to include in the list"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object",
                                "properties": {
                                    "id": {
                                        "type": "string",
                                        "example": "6055a0ae80e90e08641e22ef"
                                    },
                                    "username": {
                                        "type": "string",
                                        "example": "Layla"
                                    },
                                    "active_mission_count": {
                                        "type": "int",
                                        "example": 2
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden Error"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/users/count": {
            "get": {
                "tags": [
                    "User Management"
                ],
                "summary": "User Count [ADMIN]",
                "description": "Get the count of users with filters for Admin interface",
                "parameters": [
                    {
                        "in": "query",
                        "name": "status",
                        "required": false,
                        "description": "filter for the user status",
                        "type": "array",
                        "collectionFormat": "multi",
                        "items": {
                            "type": "integer"
                        }
                    },
                    {
                        "in": "query",
                        "name": "type",
                        "type": "array",
                        "required": false,
                        "description": "filter for the user type",
                        "collectionFormat": "multi",
                        "items": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "array",
                                    "example": []
                                },
                                "type": {
                                    "type": "array",
                                    "example": [
                                        1,
                                        2
                                    ]
                                },
                                "count": {
                                    "type": "int",
                                    "example": 1
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden Error"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/users/cur_missions": {
            "get": {
                "tags": [
                    "User Management"
                ],
                "summary": "User Current Missions Retrievement",
                "description": "Retrieve all missions in which the user is currently participating actively\n\nMission Statuses(enum):\n1: CREATED\n2: ONGOING\n3: PAUSED",
                "responses": {
                    "200": {
                        "description": "List of all the current missions",
                        "schema": {
                            "type": "array",
                            "items": {
                                "properties": {
                                    "_id": {
                                        "type": "string",
                                        "example": "6055a0ae80e90e08641e22ef"
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "MissionABC"
                                    },
                                    "status": {
                                        "type": "int",
                                        "example": 1
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found - User not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/users/{user_id}/approval": {
            "put": {
                "tags": [
                    "User Management"
                ],
                "summary": "User Account Approval [ADMIN]",
                "description": "approve the user account and specify their type",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "path",
                        "description": "ID of the user to delete",
                        "required": true,
                        "type": "string",
                        "example": "6055a0ae80e90e08641e22ef"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Approval data",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "approved": {
                                    "type": "boolean",
                                    "example": true,
                                    "required": true
                                },
                                "type": {
                                    "type": "int",
                                    "example": 1,
                                    "required": false
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User account status updated successfully",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "User account status is updated successfully."
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found - User not found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    }
}